add test token as zero step
