+++ deploy/000_deploy_test_token.ts
@@ -1,0 +1,39 @@
+import { DeployFunction } from 'hardhat-deploy/types';
+import { developmentChains } from '../helper-hardhat-config';
+import verify from '../utils/verify';
+
+const func: DeployFunction = async function ({ deployments, getNamedAccounts, network }) {
+  const { deploy, log } = deployments;
+  const { deployer } = await getNamedAccounts();
+
+  let token = null;
+
+  if (developmentChains.includes(network.name)) {
+    const argsToken = ['TEST', 'TST', '1249989122910552325012092'];
+
+    token = await deploy('TestToken', {
+      from: deployer,
+      args: argsToken,
+      log: true,
+    });
+  }
+
+  if (network.name == 'testnet' || network.name == 'sepolia') {
+    const argsToken = ['gBZZ', 'gBZZ', '1250000000000000000000000'];
+    token = await deploy('TestToken', {
+      from: deployer,
+      args: argsToken,
+      log: true,
+    });
+
+    if (process.env.TESTNET_ETHERSCAN_KEY) {
+      console.log('Verifying...');
+      await verify(token.address, argsToken);
+    }
+  }
+  log('----------------------------------------------------');
+
+};
+
+export default func;
+func.tags = ['main', 'testToken', 'contracts'];
